feat : 반려식물 저장 repository 생성

+++ src/main/java/com/planti/domain/userplant/dto/request/UserPlantRequestDto.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class UserPlantRequestDto {
 
     @NotNull
-    private Integer plantId;       // plant_id
+    private long plantId;       // plant_id
 
     @NotNull
     private String plantName;      // 사용자 입력 이름
